data_ingestion/requirement.txt
pydantic==2.10.5
pandas==2.2.3
faker==33.3.1
kafka-python 
black
pyarrow
boto3
cassandra-driver 
python-dotenv
snowflake-connector-python
pytest 
pytest-mock
pytest-asyncio